+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 - TomTom: Fixed auto-waypoint settings
 - Companion: Tooltips for solves/crates are now automatically refreshed
+- Fixed a lua error caused by double right click surveying in combat
 - Slight performance improvement
 - Behind the scenes changes: the addon is getting refactored to speed up development