+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 - Fixed a lua error caused by double right click surveying in combat
 - Slight performance improvement
 - Behind the scenes changes: the addon is getting refactored to speed up development
+- **9.0.3.1** Fix version in toc
+- **9.0.3.1** Fix artifact bars positioned outside of the main window
+- **9.0.3.1** Show races in the main window even if they are not discovered by the player yet
 
 ### 9.0.2